Restructured Code Base, Updated Adversary Decision Making
Structure of the code base is changed. The game map is now an immutable structure, with many branching paths out of it, called MapActions. This is better visualised as a graph, where each node is the state of the immutable game map and the edges are all possible actions choices of the current unit. Each MapAction can be queries to get the next map, or each GameMap can perform DoAction(action) to move to the next map, where it returns the new state of the game upon a valid action applied.